added teamsiteId and channelFolderId under Tab Context interface
