+++ src/public/interfaces.ts
@@ -423,6 +423,16 @@ export interface Context {
    * Where the user prefers the file to be opened from by default during file open
    */
   userFileOpenPreference?: FileOpenPreference;
+
+  /**
+   * Teamsite ID
+   */
+  teamSiteId?: string;
+
+  /**
+   * channel folder ID
+   */
+  channelFolderId?: string;
 }
 
 export interface DeepLinkParameters {
